bind account creation details

+++ Risk2/Data/Models/Account.cs
@@ -17,7 +17,7 @@ public class Account
     [Unique]
     public string AccountName { get; set; } = string.Empty;
 
-    public int InitialBalance { get; set; } = 0;
+    public double InitialBalance { get; set; } = 0;
 
     public double Risk { get; set; } = 0.0;
